Launch Gradio by default
