+++ frontend_service/__init__.py
@@ -1,1 +1,2 @@
 # Frontend service package
+from .interface import launch_gradio
